Adds Order and LineItem models
* Updated tests
* Moved `update_price_and_totals` from Schema to Model
